plotRSA(): omit K1 and K2 when there is no RSA object
Omit error message: When user does not hand an RSA object but specifies the parameters in plotRSA() and calls the "CL" or "RRCL" model, the lines K1 and K2 cannot be plotted because they require application of clRange which requires an RSA object. Exclude this case to avoid a respective error message.