plot CL or RRCL: alpha for K1 and K2
New option claxes.alpha in plot() function. When plotting the "CL" or "RRCL" model with the axes K1 and K2, this option allows users to change the alpha level for which the regions of significance (i.e., the positions of the lines K1 and K2 that demarcate these regions) are computed.